新增订单

+++ art-museum-manage/src/main/java/com/lz/manage/model/dto/goods/GoodsInsert.java
@@ -1,11 +1,12 @@
 package com.lz.manage.model.dto.goods;
 
-import java.io.Serializable;
-import java.util.Date;
-import com.fasterxml.jackson.annotation.JsonFormat;
+import com.lz.manage.model.domain.Goods;
 import lombok.Data;
 import org.springframework.beans.BeanUtils;
-import com.lz.manage.model.domain.Goods;
+
+import java.io.Serializable;
+import java.math.BigDecimal;
+
 /**
  * 商品信息Vo对象 tb_goods
  *
@@ -13,35 +14,52 @@ import com.lz.manage.model.domain.Goods;
  * @date 2026-02-09
  */
 @Data
-public class GoodsInsert implements Serializable
-{
+public class GoodsInsert implements Serializable {
     private static final long serialVersionUID = 1L;
 
-    /** 关联藏品 */
+    /**
+     * 关联藏品
+     */
     private Long collectionId;
 
-    /** 名称 */
+    /**
+     * 名称
+     */
     private String name;
 
-    /** 价格 */
-    private Long price;
+    /**
+     * 价格
+     */
+    private BigDecimal price;
 
-    /** 销量 */
+    /**
+     * 销量
+     */
     private Long sales;
 
-    /** 状态 */
+    /**
+     * 状态
+     */
     private String status;
 
-    /** 库存 */
+    /**
+     * 库存
+     */
     private Long inventory;
 
-    /** 主图 */
+    /**
+     * 主图
+     */
     private String imageSrc;
 
-    /** 描述 */
+    /**
+     * 描述
+     */
     private String description;
 
-    /** 备注 */
+    /**
+     * 备注
+     */
     private String remark;
 
     /**
